Drop electedness check in TransportClusterUpdateSettingsAction
`TransportClusterUpdateSettingsAction` checks whether it's still the
elected master before submitting the followup reroute task. Such a check
can only be a best-effort thing because we might stand down as elected
master after the check, but before the task is executed. Morever it's
not necessary: if skipped, the reroute task will fail with a
`NotMasterException` yielding the same API response either way.

This commit removes the unnecessary electedness check.