Use a highly accurate TDigest to compute MAD.
MedianAbsoluteDeviation builds a second TDigest over the centroids of
the TDigest that's computed over the input. The second TDigest should
have only a handful of values, so using an AVLTreeDigest produces more
accurate results with no significant performance penalty.

Note that the HybridDigest can't use SortingDigest internally for large
inputs, even if the number of centroids is fairly small (likely less
than 1000).