Remove incorrect check for shard-in-recovery in IndicesClusterStateService
This check was only there to fix the test that this PR adjust slightly.
Very rarely, we would not get the translog corrupted exception in the
test as a result of the engine failure happening before the recovery
failure. I just made it so we don't assert anything for this rare case.