Lookup enrich queries in input order (ESQL-1324)
We currently have the enrich lookup implemented in sorted mode. However,
we should have started with an unsorted mode, which performs the lookup
in the order of the input queries, for these reasons:

- Enrich indices are typically small, containing fewer than 10K
documents. With such small indices, the unsorted mode should outperform
the sorted mode as it eliminates the need for three sorting operations.

- The range match type does not benefit much from sorting queries.

This pull request removes the sorted mode and replaces it with the
unsorted mode. Also, this PR adds support for the range match type for
fields such as ip_range, date_range, and long_range. After the feature
freeze, I will reintroduce the sorted mode as an optimization
specifically for large enrich indices.