Support rewriteAndFetch on the shard in `DfsQueryPhase` (#97152)
The following phases already support `rewriteAndFetch` on the shard:  -
The `DfsPhase` for gathering term statistics  - QueryThenFetch phase

It seems like an oversight, and honestly prevents more complex rescorers
(like the inference_rescorer) from working in the DFSQuery phase.

This commit adds rewriteAndFetch for the `DfsQuery` phase to bring it
into parity with the other document shard query phases.