Expose SettingsUpdater (#97163)
Today the `SettingsUpdater` is a package-private class, visible only to
the update-settings API action framework, but this functionality is more
generally useful. For instance the `ReservedClusterSettingsAction` needs
it, but can only access it by creating an update-settings task which it
then runs directly (outside of the master service). This is kind of
backwards: the update-settings task should be isolated to the API which
uses it, and the general-purpose `SettingsUpdater` should be more widely
accessible.

This commit moves `SettingsUpdater` to `o.e.common.settings`, adjusts
`ReservedClusterSettingsAction` to use it, and makes the update-settings
task private.