Use collector managers in DfsPhase (#96689)
This change switches the part of the DfsPhase that runs a knn query
from a single thread collector to a potentially multi-threaded collector
manager and optional profiling wrapper. This doesn't in itself enable
concurrent execution but lays the foundation for a switch of the executor
in the ContextIndexSearcher to actually enable this.