Allow uses to run the painless execute API on a remote cluster shard.
We added a clusterAlias to the Painless execute Request object, so that index
expressions in the request of the form "myremote:myindex" will be parsed to
set clusterAlias to myremote and the index to "myindex".

If clusterAlias is null, then it is executed against a shard on the local cluster, as before.
If clusterAlias is non-null, then the SingleShardTransportAction is sent to the remote cluster,
where it will run the full request (doing remote coordination). Note that before forwarding
the request to the remote cluster, we nullify the clusterAlias in the request, so that the
remote coordinator will interpret the request to run it locally, rather than try to lookup
a remote cluster.

Closes #92126 and kibana issue #145861