Remove service account trace log of bearer tokens
The service account is the first in the authentication chain.
When `ServiceAccountToken` logger is set to `TRACE`, it would
log all bearer tokens (even `OAuth2` and `JWT` tokens).

While this log might have been useful during development, there is no
real benefit having it in production as it's not clear to users that
this could unintentionally leak valid credentials.