Remove redundant read of the data stream write index (#97277)
This code works on one cluster state version. We read the `DataStream`
from  the same cluster state twice. The write index will always be the
same in this case. (as we're not reading the `DataStream` from a new
cluster state version)

This PR drops a redundant (and confusing) reference the same write index
of the data  stream.