[ML] Evolve use of versions in decision making
In this first pass at evolving the way the ML and Transform plugins use version numbers, where possible TransportVersions are used to answer questions like "is the cluster in such a state to support a particular feature".

Initally creating as a draft PR to see how it affects the BWC tests.