Fix inefficient ActionListener.delegateFailure and ActionListener.wrap usage
ActionListener.delegateFailure is not the optimal choice when the response handler
always invokes delegate.onResponse, it's just needlessly verbose, error-prone and sometimes
less efficient when multiple map/delegation calls are chained.
ActionListener.wrap is just a less efficient version of ActionListener.delegateAndWrap when
the error handler is `listener::onFailure` -> fixup a couple of these as well.