Mark `apm_user` for removal in "a future version" (#87674)
Meta issue: elastic/kibana#116760

The `apm_user` role was marked as deprecated in 7.13 and was supposed to be removed in 8.0 but it
didn't happen. Now we are aiming to remove the role in 9.0 and are updating the deprecation message.

All mentions of `apm_user` role have been removed from docs and in-product mentions
in elastic/kibana#132790.