ProfileCollectorManager to support child profiler collector
ProfileCollectorManager is our collector manager used for profiling
which supports parallel collection. We have integrated it in the DFS
phase which has a very simple collection hierarcy: a top docs collector
only.

In the query phase, we may have a top-level collector that holds two sub
collectors where each one is profiled separately.

This commit adds support for children profilers to
ProfileCollectorManager so that we can use it in the query phase as
well.