Avoid wrapping leaf collector when possible in QueryPhaseCollector
The recently introduced QueryPhaseCollector introduces some overhead
caused by wrapping of the leaf collector that could be avoided.

When either a post_filter, or a min_score, or terminate_after are
provided, we have no other choice than wrapping to filter collection,
but if none of these are used, we can skip wrapping when one of the two
leaf collectors early terminates.