Limit several determinization limits for automata
Determinization of automata can be an expensive operation, which is why Lucenes
MinimizationOperations allows to limit the amount of work allowed in this step.
The default limit should be Operations.DEFAULT_DETERMINIZE_WORK_LIMIT, but we
use Integer.MAX in several places. Instead, we can limit this to even 0 in many
places where by the construction of the automaton it is clear that it is already
determinized.
This changes this limit to 0 where possible and adds tests and assertions to
make sure the constructed automata are deterministic already, also simplifying
some internal API along the way.

Relates to #96572