[Profiling] Propagate mget failures (#97451) (#97483)
The Universal Profiling plugin uses the `mget` API internally to enrich
profiling data. If the `mget` API fails to retrieve an item, it will not
error but instead provide error details per item. Previously we have
skipped such items leading to puzzling results e.g. when a user was
lacking proper permissions (user would see an empty page instead of an
error). With this commit we abort processing on the first failed item
and return an appropriate error response.