[Profiling] Fix off-by-one error in test assert
The test case `testRandomLengthListGreaterThanSliceCount` is supposed to
test behavior of a list slice implementation when the input list
contains more elements than desired slices. However, there was an
off-by-one error so sometimes the input list was exactly the same size
as the desired number of slices and thus the assertion was wrong. With
this commit we ensure that the input list is always larger than the
desired number of slices.