Introduce a collector manager for PartialHitCountCollector
When size is set to 0, a PartialHitCountCollector is used to collect
total hits and optionally early terminate the total hits counting. This
commit introduces a collector manager for such collector so that its
collection can be parallelized across slices.

Note that this commit does not wire the new collector manager yet in the
query phase. That will happen in a follow-up change.