Avoid restarting `MonitoringService` in tests
In #97553 we will forbid moving a lifecycle component from `STOPPED` to
`STARTED`, because we don't ever do this in production and there are
several places where to do so would be a bug. One of the few places we
do this transition today is in `MonitoringIntegTestCase`. This commit
replaces the lifecycle transitions in that test suite with a
pause/unpause mechanism that does the same thing to the underlying
service.