[ML] fix weird change_point bug where all data values are equivalent (#97588)
If a user calls change_point aggregation and all the bucket values are exactly the same, we may run into weird floating point errors when calculating statistics. When we have no variance and standard deviation is 0, we should indicate that there is no change point and its a stationary set of data.