Readjust plans again - now I embed the SearchShards API test in the existing tests
and pattern the asserts as close as possible to the Search API asserts.
In doing this I found that the only part that seems to continue to be flaky
are the expectThrows - occassionally these do not throw, but return results,
so I allow it to either throw or if can get results in those cases none
of the shards should be marked as skipped since the snapshot should be
(mostly) unavailable.