Move slow FieldData#toString(...) logic to where it is needed (#97438)
Change FieldData#toString(...) to delegate to SortedSetDocValues#docValueCount() instead of computing it in the advanceExact(...) method.

In some cases it is needed to pre-compute doc value count, because the number of ordinals is dynamic. For example in the case of flattened field. This logic is moved to KeyedFlattenedLeafFieldData, the only place where it is needed.

As part of profiling downsampling, advancing twice showed up as part of rolling up label fields.
However for almost all these label fields, pre-computing doc value count isn't needed.

Relates to #97141