Max headrooms should not apply if a watermark set
This is a solution for clusters that upgrade to 8.5.0+ that have
only one watermark set, and would otherwise fail to upgrade because
of an exception thrown.

However, it may break any existing 8.5.0+ clusters that may have
just one max headroom set.

So it is a double-edge sword, and opening this for discussion.

Fixes #93468