Disable MMAPed shared cache file IO in Windows tests (#97639)
We don't have code to umap the shared bytes mmap so it only gets
unmapped after GC on OSX and Linux and after JVM exit on Windows. This
is not an issue for tests on OSX and Linux but on Windows this trips
assertions. There's no use-case for making this work on Windows for now
so lets just disable using mmap in tests on Windows.

closes #97626 closes #97638