Introduce a collector manager for QueryPhaseCollector (#97410)
The query phase now relies on QueryPhaseCollector to wrap top docs and aggs collector as well as to apply terminate_after, post_filter and min_score. In order to enable inter-segment concurrency, we need to introduce a corresponding collector manager that can create multiple query phase collectors and reduce them to a single resultset.

Note: we are only introducing the collector manager and corresponding tests, but not using it yet in QueryPhase.