GEO_POINT for ESQL
At this point we support:
 * display of geo_point types in results tables
 * ToGeoPoint function which takes either ane encoded long or a WKT keyword
 * Support for geo_point as argument to ToString (renders as WKT)
 * Support for GeoCentroid() aggregating function
 * Support intersects function for point-in-polygon predicates

Also, the support for GeoPoint is as an encoded long internally, so we do not support a new Block type in this prototype.

There was no need for the new spatial types to be listed as supported for all query languages.
Also, I noticed that once supported the tests for geo would fail on a missing `@timestamp` field for EQL,
since that seems to plan an implicit sort by time, so I added that to the mapping-geo.json file so that future geo support is not surprised by this unexpected error.

We make use of Lucene Component2D capabilities for point-in-polygon,
but could instead use the xpack.spatial module for much more complex geometry relationship comparisons, but that requires adding a dependency.