Change pow() behaviour with unsigned long to return long
The original design was that floating types returned floating types, and non-floating returned non-floating, similar to the behaviour of MQ-SQL. However, the unsigned long support deviated from this by returning double. This change returns to the original behaviour.

One negative of this is that we can no longer return Infinity for cases of numerical overruns.