Add completion_time time field to async_search get and status response
The completion_time is set as the start_time (already present) plus the 'took'
time that is set in the SearchResponse object and only if the isRunning status == false
since took is set even for in-progress searches.

We use the 'took' field because it is based on relative time, not absolute wall clock time
which can go backwards due to NPT issues. See the comments in TransportSearchAction about
the SearchTimeProvider for details.

Closes #88640