Remove locking on shared blob cache service for hot pages (#97674)
Locking on the cache service is unnecessary, the access time can be
safely checked through a volatile read. If another thread concurrently
promotes the entry, then the time check under the lock will practically
always fail and without a concurrent update to the entry it doesn't
matter in the first place.
-> This removes the non-scalable (in number of threads) global lock for
pages that are hot most of the time, which should do away with some tail
latency that we see in benchmarks.