[Profiling] Tighten index name check
All profiling-related indices contain a version identifier in their
name. In some places we need to check whether an index exists regardless
of its version. Previously this check was too lenient because it also
considered indices with the same prefix (e.g. a check for
`profiling-stackframes` would also match `profiling-stackframes-all`).
With this commit we tighten the check to match the full name until the
version identifier.