Release master service task on timeout
Today when a task in a master service queue times out it remains in the
queue until its batch is eventually processed. This might mean we retain
a reference to the task (and therefore any dependent listeners and other
objects) for an extended period of time after it has failed. This commit
adjusts the timeout handling to drop the reference to the task when it
completes.