Wire QueryPhaseCollectorManager into the query phase
We recently introduced a collector manager for QueryPhaseCollector.
This commit wires it into the query phase, which requires also switching
the query phase to using ProfileCollectorManager which supports
concurrent profiling.

For simplicity, the wrapping for the aggs collector manager when
profiling is on is moved to the query phase, as it removed the need for
casting when creating a ProfileCollectorManager with children.

Note: there is no concurrency enabled with this as we don't yet set the
executor to the IndexSearcher