Track ongoing tasks in TaskExecutionTimeTrackingEsThreadPoolExecutor (#97661)
For Stateless autoscaling, to be able to account for ongoing tasks in
the estimated ingestion load, we need to keep track of the ongoing
tasks. This is important because currently the only stats that we get
from the executor is the `TotalTaskExecutionTime` which is updated only
once a task is finished. This can lead to cases where if we do have long
tasks (write tasks) that last longer than a sampling period (1s), we'd
observe no write load and assume there is no load, because the tasks
would be included in the stats only once they are finished. Recording
ongoing tasks allows taking these into consideration.

Relates ES-6391