Improve how connections are rejected after shutdown
Prior to this PR, when a channel was accepted after Elasticsearch
received a shutdown signal, `AbstractHttpServerTransport` would
directly close the connection. This PR makes it throw an exception
instead, which results in the connection being closed by Netty.
This keeps the flow of control more consistent and readable, as
well as logging additional details such as the host which made
the request.