Remove transport version from main endpoint response
This is a partial backport of #97675.

The transport version is a low level detail of how a node behaves in a cluster. It was recently added to the main endpoint response, but is too low level and should be moved to another endpoint TBD. This commit removes it from the main response, though keeps it as part of the wire format for the MainResponse to avoid any need for a bwc change to the format.