Increase driver iteration time limit (ESQL-1445)
Today, we pause the execution of a Driver and yield the executing thread
to another Driver in the following cases:

1. When the downstream or upstream pipeline is not fast enough. For
example, if the coordinator hasn't proceeded with all the output pages
from the data-node, the data-nodes can yield their executing threads.
This is done via the buffer of the exchanges.
2. When the number of iterations exceeds the limit (defaults to 10,000).
3. When the execution time exceeds the limit (defaults to 200ms).

I don't see the benefits of the second and third scenarios except when
dealing with slow queries. Throttling slow queries can provide
opportunities for fast queries to complete quickly. However, this
increases memory usage, as more outstanding requests are executed
simultaneously, instead of being limited to the number of threads.
Rescheduling drivers can also be expensive, as it involves
re-initializing certain Lucene structures.

In this pull request, we increase the time limit of a Driver iteration
from 200ms to 5 minutes.