Allow more parts of Build to be filled in
Build represents information about the running build of ES. It has
some methods that assume some structure to the existing version.
Additionally MainResponse contains Version just to be able to retrieve
the min compat versions.

This commit makes all of these bits of info part of the state of Build.
It also removes Version from MainResponse as it is no longer necessary.