Cache latest file region in SharedBlobCacheService.CacheFile
In many cases we access the same file in rapid succession.
We can cache the most recently accessed region in the file
to avoid the costly lookup from the CHM.
We only do this for a time window that is inside the throttle
time for frequency counting in the LFU logic to avoid the overhead
of randomly jumping to a synchronized method on the fast-path.