Improve exception handling in Coordinator#publish
We shouldn't complete the publication listener under the `mutex`, and
it's a little tricky to see which exceptions are or are not impossible
from the way the code is written today.

This commit solves all that by throwing the exception out to a handler
outside the mutex, and splitting the enclosing `try...catch` block into
several smaller blocks.

It also fixes a bug where we could fail to complete the publish listener
if the publication failed to start properly. With this change, we now
clearly complete the publication listener on all failure paths.

Closes #97798