[ML] Fix ML usage action when ML sub-features are disabled (#97820)
To support differentiation of serverless product types it's now
possible to disable subsets of ML functionality.

The ML usage action was not updated to reflect this, so still
attempted to call actions to find out about usage of disabled
sub-features. Because that action is designed to never fail,
it actually returned reasonable output in these cases. However,
it generated a lot of log spam with warnings about calling
non-existent actions.

This PR adjusts the logic of the ML usage action to take into
account which sub-features might be disabled, and avoid calling
actions which won't exist in this case.