Remove redundant ref-counting in SharedBytes.IO
No need to use ref-counting here to keep the shared bytes alive
after they got closed. SharedBytes live for as long as the node lives,
so if this ever becomes an issues we should fix it by fixing the
shutdown order of services and not deal with the significant overhead
of ref-counting on every read.