Allow "slow" MethodHandle invocations in tests
MethodHandles provide an alternative to reflective method invocation. In
many cases they are faster than reflective methods. But sometimes they
might be just as slow. We currently forbid use of these "slow" variants,
even in tests. But tests don't really care about speed, per se, and they
are no worse than using reflection, which is what we would resort to in
tests.

This commit adjusts the forbidden rules to disallow "slow" MethodHandle
invocations only in server code, not in tests.