Add test cluster setting to control the minimum number of docs in a concurrent slice (#97740)
The current slicing algorithm for concurrent search only creates multiple slices iff it can add at least 50k documents in each slice. This means that we need to have at least 100K documents in order to trigger concurrent search.

In tests, we can create several segments, but normally those segments contains few documents, therefore they will never trigger concurrent search.

This PR adds a cluster setting that can control the minimum number documents we require to create multiple slices. This setting is only registered on tests, therefore it is not exposed to the users. During test we can set this number to a
very low value to increase the likelihood that the test will exercise the concurrent code-path.

This change affects ESIntegTestCase as well as EsSingleNodeTestCase.

---------

Co-authored-by: Luca Cavanna <javanna@apache.org>