Extract cold path from FrozenIndexInput
Just like we did in the search index input, we should have the fast-path in a small method
so it inlines (hopefully) in scenarios like reading backwards in a file where this method is very hot.
Also, profiling showed that allocating the boxed numbers for the trace call is a visible contributor to the
cost of this method, so wrapping that call in a manual check.