[Profiling] Consider static settings in status
Previously the profiling status check only considered the dynamically
set value for `xpack.profiling.templates.enabled`. If a user set a
value statically in `elasticsearch.yml` the status check did not reflect
that leading to puzzling responses. With this commit we check not only
dynamically defined cluster settings but also the static ones before
returning the setting's default value.