Fix error in csv tests (ESQL-1474)
Surprise! There was an off by one error in the CSV tests which made it
so the tests would pass if we specified exactly one fewer row than was
returned. More than one fewer row would fail. This corrects the error
and tries to make the code a little more "normal java" with the index
variable of the `for` loop not escaping. Escaping index variables have a
defined behavior, but I totally didn't see the off by one error until I
was testing something else. But there it was.

Anyway! This fixes the error and fixes the tests.