Ensure frozen indices have correct tier preference
We've seen a series of cases where partially mounted indices either had
an invalid tier preferece (e.g. `data_content`) or a missing tier
preference.

This tightens up our safety nets to make sure that when we change the
index configuration (e.g. we remove the tier preference using the update
settings API), migrate a deployment to data tiers, restore an index, or upgrade
the persisted metadata we set the tier preference to `data_frozen` for
partially mounted indices.