Flatten class hierarchy for searchable snapshot index inputs
There's not much to gain from having a shared
BaseSearchableSnapshotIndexInput except for a trivial couple of fields.
On the other hand, with upcoming improvements to the shared blob cache
service, we don't want `FrozenIndexInput` to still be a
`BufferedIndexInput`.
This PR is one step towards that end, that flattens the class hierarchy
by one. A subsequent step would remove the inheritance of `BufferedIndexInput`
from `MetadataCachingIndexInput` to enable FrozenIndexInput to directly
read from mmaped buffers in the shared bytes cache (if mmap io is used).