Fork remote-cluster response handling (#97922)
Today all responses to remote-cluster requests are deserialized and
handled on the transport worker thread. Some of these responses can be
sizeable, so with this commit we add the facility for callers to specify
a different executor to handle this work. It also adjusts several
callers to use more appropriate threads, including:

- responses from CCR-related admin actions are handled on `ccr`
- responses from field caps actions are handled on `search_coordination`