fix: hash the tsid if required
A Lucene limitation on doc values for UTF-8 fields does not
allow us to write string fields whose size is larger then
32K. This limits our ability to map more than a certain
number of dimension fields for time series indices.

To overcome this limitation we hash the tsid if the size of
the fields exceeds the limit and prepend the hash with a prefix
that makes it easier to query fields and investigate how many
documents are using an hahed tsid.