Make some code on SharedBlobCacheService hot path more compact
Make some code here a little friendlier to inline on the very hot path
by removing assertions and safety checks (that are redundant anyway).
This doesn't matter massively now but the closer we move to directly
accessing the mmap in the SharedBytes.IO the more we are hurt by any
extra code/checks.
Also, replace the evicted AtomicBoolean with a hand-crafted volatile
+ var handle solution to save some memory and indirection as we look
into potentially making the cache more fine-grained.