Expand search timeout tests (#97945)
We don't have a lot of coverage for search timeout support. There is an existing integration test, but it only checks the timed_out flag returned in the search response. It should also check that partial results are properly returned.

Also, this commit adds unit tests for the timeout support at the query phase level, which are not dependent on timing.