Run deleting old index gens after snapshot in the background (#98062)
Bit of a strange way to deal with the test failure but I think this is a
valid optimisation in isolation. I think it's fine to have that one
tasks forked and executing without waiting for it to the snapshot pool. 

There's no need to run the old index gens cleanup before completing a
snapshot operation (and starting the next operation). This also makes it
so the snapshot listener is resolved on the master thread, so we can get
a consistent view of the cluster state when calling
clusterService.state() in the callback. This fixes test failure #97572
which resulted from accessing the state on the `SNAPSHOT` pool after a
master failover.

Fixes #97572