Move instructions about new transport versions
Today the instructions about adding a new transport version are in the
class javadoc, but these docs are really more for the users of a class
than its maintainers, and it's easy to miss that there are vital
instructions here.

This commit moves them somewhere a little less missable, and clarifies
that the list of versions should be _only_ versions, no comments or
whitespace.