Prepare parallel collection in Dfs phase (#97416)
With the infrastructure for parallel collection in place, this change prepares
enabling it when creating the SearchContext in the SearchService. The current
decision logic takes into account a new global dynamic setting
("search.concurrency_enabled") that is on by default but can be used to turn off
parallelism if it is problematic for some reason.
Currently, parallel collection is only enabled for the Dfs phase with future
extensions possible by modifying the decision logic in SearchService, which
currently can take into account phase and content in the SearchSourceBuilder.
Enabeling parallel collection is pending on the introduction of a separate
threadpool which needs to be used in the DefaultSearchService constructor,
setting an executor for the ContextIndexSearcher that manages the parallel
collection.