Cleanup deleting of old root level repository metadata
Test #98082 seems to be failing because waiting for no running tasks on
the snapshot threadpool isn't 100% safe (the active count may be off).
This fixes the issue of a background deletion by moving the deletion of
old root level blobs to the code path that already does async deletion
of shard level metadata blobs.
There was never any reason for these deletes not to be covered by that
logic in the first place.