Change how grouping aggs track null (ESQL-1328)
This changes how grouping aggs track `null` to fix bugs in the old
tracking caused by `null`s arriving after non-null values and things
like that. Now there are two "modes" from the `X-ArrayState` classes:
1. A mode where we do not track null and rely on the `selected` list of
buckets produced by the `BlockHash`. This mode is appropriate right up
until you see the first `null` value.
2. A mode where we do track which values are null and can reply with
`null` even for values `BlockHash` selects if we've only ever added
`null` values to that slot. This is appropriate for all data, but less
efficient so we only transition to it when we receive our first `Block`
containing `null` values.

The transition is *interesting* because we need to know which values
aren't `null`. Luckily, `BlockHash` has that information. Usually it's
just "everything in this range" but for `boolean` values it isn't. And,
I expect, for ordinals it won't be either. So we just ask `BlockHash` at
the moment of the transition. That required changing the interface
around a little bit so we could ask *before* the group keys for the next
block was added.