Add stall_time_seconds to the error_query DSL testing query.
This allows simulating long running queries (particularly useful for CCS testing).
In addition a "NONE" exception type was added for cases where you want the stall time
greater than 0 but to not throw an Exception.

Most of the runtime logic for the Query was moved out of the ErrorQueryBuilder
and into a new ErrorQuery class (extending Lucene Query class), so that this can be
tested in conjuction with search timeout flag to induce query timeouts and allow
testing of that feature as well.